+++ src/main/scala/dk/versio/aoc2020/Dec01.scala
@@ -2,7 +2,7 @@ package dk.versio.aoc2020
 
 object Dec01 extends App {
 
-  def get2020Combinations(file: String, n: Int): Option[Int] = Common
+  def get2020Combinations(file: String, n: Int): Option[Long] = Common
     .loadNumbers(file)
     .combinations(n)
     .find(p => p.sum == 2020)
